pyproject.toml
[project]
name = "edreporter"
version = "0.1.0"
description = "Human-in-the-loop PDF article segmentation tool with OCR and DOCX export"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.39.0",
    "pymupdf>=1.24.0",
    "pillow>=10.4.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "pydantic>=2.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "black>=24.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

